chore: util fn exports
